edit: execute file 구조변경경
