+++ .vscode/settings.json
@@ -20,6 +20,10 @@
 		"readline.h": "c",
 		"__bit_reference": "c",
 		"chrono": "c",
-		"locale": "c"
+		"locale": "c",
+		"stdio.h": "c",
+		"execute.h": "c",
+		"exec_word_util.h": "c",
+		"expansion.h": "c"
 	}
 }
