+++ execute/exec_node_util.c
@@ -1,0 +1,140 @@
+
+#include "../include/minishell.h"
+#include "../include/exec_word_util.h"
+
+// void	exec_subshell(t_node *node, t_ctx *p_ctx)
+// {
+// 	int		pid;
+// 	t_node	*lhs;
+
+// 	lhs = node->left;
+// 	pid = fork();
+// 	if (pid)
+// 	{
+// 		exec_word(lhs, p_ctx);
+// 		wait_queue();
+// 	}
+// 	enqueue(pid);
+// 	wait_queue();
+// }
+
+// void	exec_or(t_node *node, t_ctx *p_ctx)
+// {
+// 	int		pid;
+// 	t_node	*lhs;
+// 	t_node	*rhs;
+
+// 	lhs = node->left;
+// 	rhs = node->right;
+// 	exec_word(lhs, p_ctx);
+// 	if (p_ctx->exit_status != 0)
+// 	{
+// 		exec_word(rhs, p_ctx);
+// 	}
+// }
+
+// void	exec_and(t_node *node, t_ctx *p_ctx)
+// {
+// 	int		pid;
+// 	t_node	*lhs;
+// 	t_node	*rhs;
+
+// 	lhs = node->left;
+// 	rhs = node->right;
+// 	exec_word(lhs, p_ctx);
+// 	if (p_ctx->exit_status == 0)
+// 	{
+// 		exec_word(rhs, p_ctx);
+// 	}
+// }
+
+// void exec_pipe(t_node *node, t_ctx *p_ctx)
+// {
+// 	t_node	*lhs;
+// 	t_node	*rhs;
+// 	int		pipe_fd[2];
+// 	t_ctx	*aux;
+
+// 	lhs = node->left;
+// 	rhs = node->right;
+// 	pipe(pipe_fd);
+// 	*aux = *p_ctx;
+// 	aux->fd_close = pipe_fd[STDIN];
+// 	aux->fd[STDOUT] = pipe_fd[STDOUT];
+// 	exec_word(lhs, aux);
+
+// 	*aux = *p_ctx;
+// 	aux->fd[STDIN] = pipe_fd[STDIN];
+// 	aux->fd_close = pipe_fd[STDOUT];
+// 	exec_word(rhs, aux);
+// }
+
+// void	exec_input(t_node *node, t_ctx *p_ctx)
+// {
+// 	t_node	*lhs;
+// 	t_node	*rhs;
+
+// 	lhs = node->left;
+// 	rhs = node->right;
+
+// 	p_ctx->fd[STDIN] = open(rhs->word.buf, O_RDONLY, 0644);
+// 	exec_word(lhs, p_ctx);
+// }
+
+// void exec_heredoc(t_node *node, t_ctx *p_ctx)
+// {
+// 	t_node	*lhs;
+// 	t_node	*rhs;
+
+// 	lhs = node->left;
+// 	rhs = node->right;
+
+// 	p_ctx->fd[STDIN] = open(rhs->word.buf, O_RDONLY, 0644);
+// 	exec_word(lhs, p_ctx);
+// }
+
+// void exec_output(t_node *node, t_ctx *p_ctx)
+// {
+// 	t_node	*lhs;
+// 	t_node	*rhs;
+
+// 	lhs = node->left;
+// 	rhs = node->right;
+
+// 	p_ctx->fd[STDOUT] = open(rhs->word.buf, O_CREAT | O_TRUNC | O_WRONLY, 0644);
+// 	exec_word(lhs, p_ctx);
+// }
+
+// void exec_append(t_node *node, t_ctx *p_ctx)
+// {
+// 	t_node	*lhs;
+// 	t_node	*rhs;
+
+// 	lhs = node->left;
+// 	rhs = node->right;
+
+// 	p_ctx->fd[STDOUT] = open(rhs->word.buf, O_CREAT | O_APPEND| O_WRONLY, 0644);
+// 	exec_word(lhs, p_ctx);
+// }
+
+void	exec_word(t_node *node, t_context *p_ctx)
+{
+	char	**argv;
+
+	// node에 저장된 cmd line argument 배열 parsing
+	argv = make_argv(node->word);
+	// 빌드인 or PATH에 경로등록 or 현재 디렉토리에 존재하는 명령
+	if (ft_strchr(argv, '/') == NULL)
+	{
+		// 빌트인
+		if (exec_builtin() == NULL) {
+			// PATH(argv에 경로를 붙혀서 실행해야하는 경우), 현재 디렉토리 search
+			search_and_fork_exec();
+		}
+	}
+	// 경로가 명시된 경우(상대경로 or 절대경로)
+	else if (can_access(argv[0], p_ctx))
+		fork_exec(argv, p_ctx);
+	set_exit_status(p_ctx->exit_status);
+	free_argv(argv);
+}
