+++ include/exec_word_util.h
@@ -1,0 +1,4 @@
+char	**make_argv(char **word_arr);
+void	free_argv(char **argv);
+void	set_exit_status(int exit_status);
+int	*get_exit_status(void);
