+++ include/execute.h
@@ -1,5 +1,23 @@
+/* ************************************************************************** */
+/*                                                                            */
+/*                                                        :::      ::::::::   */
+/*   execute.h                                          :+:      :+:    :+:   */
+/*                                                    +:+ +:+         +:+     */
+/*   By: dasong <dasong@student.42.fr>              +#+  +:+       +#+        */
+/*                                                +#+#+#+#+#+   +#+           */
+/*   Created: 2023/08/15 11:48:17 by dasong            #+#    #+#             */
+/*   Updated: 2023/08/15 11:48:20 by dasong           ###   ########.fr       */
+/*                                                                            */
+/* ************************************************************************** */
 
+#define STDIN 0
+#define STDOUT 1
 
-char	**env_to_strarr();
-int	list_len(t_list	*env_node);
+typedef struct s_context
+{
+	int	exit_status;
+	int	fd[2];
+	int	fd_close;
+	int	exit_status;
+}	t_context;
