+++ include/minishell.h
@@ -65,3 +65,5 @@ typedef struct s_tokenizer
 	char		*start;
 	char		*end;
 } t_tokenizer;
+
+t_list	**get_envp(void);
